split unshare related code into its own header file
