+++ src/process-events.c
@@ -11,6 +11,8 @@
 #include "offsets.h"
 #include "repeat.h"
 #include "common.h"
+#include "process/unshare.h"
+#include "process/helpers.h"
 
 #define CLONE_ARGS_SIZE_VER0 64 /* sizeof first published struct */
 #define CLONE_ARGS_SIZE_VER1 80 /* sizeof second published struct */
@@ -27,11 +29,6 @@
 
 // these structs are used to send data gathered/calculated in a kprobe
 // to the kretprobe.
-
-typedef struct {
-    u64 pid_tgid;
-    int unshare_flags;
-} incomplete_unshare_t;
 
 typedef struct {
     u64 pid_tgid;
@@ -73,28 +70,6 @@ struct bpf_map_def SEC("maps/percpu_path") percpu_path = {
     .namespace = "",
 };
 
-// The map where process event messages get emitted to
-struct bpf_map_def SEC("maps/process_events") process_events = {
-    .type = BPF_MAP_TYPE_PERF_EVENT_ARRAY,
-    .key_size = sizeof(u32),
-    .value_size = sizeof(u32),
-    .max_entries = 0, // let oxidebpf set it to num_cpus
-    .pinning = 0,
-    .namespace = "",
-};
-
-// A map of unshares that have started (a kprobe) but are yet to finish
-// (the kretprobe).
-struct bpf_map_def SEC("maps/incomplete_unshare") incomplete_unshares = {
-    .type = BPF_MAP_TYPE_LRU_HASH,
-    .key_size = sizeof(u64),
-    .value_size = sizeof(incomplete_unshare_t),
-    // this is lower than exec or clone because we don't foresee that many concurrent unshares
-    .max_entries = 256,
-    .pinning = 0,
-    .namespace = "",
-};
-
 // A map of clones that have started (a kprobe) but are yet to finish
 // (the kretprobe).
 struct bpf_map_def SEC("maps/incomplete_clone") incomplete_clones = {
@@ -126,86 +101,6 @@ struct bpf_map_def SEC("maps/buffers") buffers = {
     .namespace = "",
 };
 
-typedef struct
-{
-    process_message_warning_t code;
-    error_info_t info;
-} local_warning_t;
-
-// A "cpu local" warning so we can easily get/set the current warning
-// at any point
-struct bpf_map_def SEC("maps/warning") percpu_warning = {
-    .type = BPF_MAP_TYPE_PERCPU_ARRAY,
-    .key_size = sizeof(u32),
-    .value_size = sizeof(local_warning_t),
-    .max_entries = 1,
-    .pinning = 0,
-    .namespace = "",
-};
-
-static __always_inline int set_local_warning(process_message_warning_t code, error_info_t info)
-{
-    local_warning_t warning = {0};
-    warning.code = code;
-    warning.info = info;
-    u32 key = 0;
-    return bpf_map_update_elem(&percpu_warning, &key, &warning, BPF_ANY);
-}
-
-/// Returns the offset to a field. If not found the current CPU's
-/// warning_info is set and NULL is returned
-static __always_inline u32* get_offset(u64 field) {
-    u32 *offset = (u32 *)bpf_map_lookup_elem(&offsets, &field);
-    if (offset == NULL) {
-        error_info_t info = {0};
-        info.offset_crc = field;
-        set_local_warning(PMW_READING_FIELD, info);
-    }
-
-    return offset;
-}
-
-/// Equivalent to &(base->field). If not found the current CPU's
-/// warning_info is set and NULL is returned
-static __always_inline void* ptr_to_field(void *base, u64 field) {
-    u32 *offset = get_offset(field);
-    if (offset == NULL) return NULL;
-
-    return base + *offset;
-}
-
-/// Equivalent to *dst = base->field. Returns 0 on success, or
-/// negative and sets the current CPU's warning_info in case of
-/// failure.
-static __always_inline int read_field(void *base, u64 field, void *dst, size_t size) {
-    void *ptr_to_ptr = ptr_to_field(base, field);
-    if (ptr_to_ptr == NULL) return -1;
-
-    int ret = bpf_probe_read(dst, size, ptr_to_ptr);
-    if (ret < 0) {
-        error_info_t info = {0};
-        info.offset_crc = field;
-        set_local_warning(PMW_PTR_FIELD_READ, info);
-    }
-
-    return ret;
-}
-
-/// Equivalent to void base->field. If not found or the pointer is
-/// NULL the current CPU's warning_info is set and NULL is returned
-static __always_inline void* read_field_ptr(void *base, u64 field) {
-    void *dst = NULL;
-    if (read_field(base, field, &dst, sizeof(dst)) < 0) return NULL;
-
-    if (dst == NULL) {
-        error_info_t info = {0};
-        info.offset_crc = field;
-        set_local_warning(PMW_NULL_FIELD, info);
-    }
-
-    return dst;
-}
-
 static __always_inline void init_cached_path(cached_path_t *cached_path, void *path)
 {
     // cached_path->next_dentry = path->dentry;
@@ -214,13 +109,6 @@ static __always_inline void init_cached_path(cached_path_t *cached_path, void *p
     // cached_path->vfsmount = path->mnt;
     cached_path->vfsmount = read_field_ptr(path, CRC_PATH_MNT);
 }
-
-#define load_event(map, key, ty)                            \
-    void *__eventp = bpf_map_lookup_elem(&map, &key);       \
-    if (__eventp == NULL) goto NoEvent;                     \
-    ty event = {0};                                         \
-    __builtin_memcpy(&event, (void *)__eventp, sizeof(ty)); \
-    if (event.pid_tgid != key) goto EventMismatch;
 
 SEC("kprobe/do_mount")
 int kprobe__do_mount(struct pt_regs *ctx)
@@ -259,51 +147,6 @@ static __always_inline void write_null_char(buf_t *buffer)
     // "overflow".
     buffer->buf[*offset & (MAX_PERCPU_BUFFER - 1)] = '\0';
     *offset = *offset + 1;
-}
-
-// returns NULL if offsets have not yet been loaded
-static __always_inline void *offset_loaded()
-{
-    u64 offset = CRC_LOADED;
-    return bpf_map_lookup_elem(&offsets, &offset);
-}
-
-// accepts a pointer to a `task_struct`. Returns 0 if it is not
-// user_process; 1 if it is, and -1 if an error occured.
-static __always_inline int is_user_process(void *ts)
-{
-    void *mmptr = NULL;
-    if (read_field(ts, CRC_TASK_STRUCT_MM, &mmptr, sizeof(mmptr)) < 0) {
-        return -1;
-    }
-    return mmptr != NULL;
-}
-
-// pushes a message to the process_events perfmap for the current CPU.
-static __always_inline int push_message(struct pt_regs *ctx, pprocess_message_t ev)
-{
-    return bpf_perf_event_output(ctx, &process_events, BPF_F_CURRENT_CPU, ev, sizeof(*ev));
-}
-
-// pushes a warning to the process_events perfmap for the current CPU.
-static __always_inline int push_warning(struct pt_regs *ctx, pprocess_message_t pm,
-                                        process_message_type_t pm_type)
-{
-    pm->type = PM_WARNING;
-
-    u32 key = 0;
-    local_warning_t *warning = (local_warning_t *)bpf_map_lookup_elem(&percpu_warning, &key);
-    if (warning != NULL) {
-        pm->u.warning_info.code = warning->code;
-        pm->u.warning_info.info = warning->info;
-        // reset it so we don't accidentally re-use the same code/info in a new warning
-        *warning = (local_warning_t){0};
-    }
-
-    pm->u.warning_info.pid_tgid = bpf_get_current_pid_tgid();
-    pm->u.warning_info.message_type = pm_type;
-
-    return push_message(ctx, pm);
 }
 
 // pushes a message with an extra `dynamic_size` number of bytes. It
@@ -352,56 +195,6 @@ static __always_inline int write_string(const char *string, buf_t *buffer, const
         *offset = *offset + sz;
         return sz;
     }
-}
-
-// fills the syscall_info with all the common values. Returns 1 if the
-// task struct is not a user process or if the offsets have not yet
-// been loaded.
-static __always_inline int fill_syscall(syscall_info_t *syscall_info, void *ts, u32 pid)
-{
-    if (!offset_loaded())
-        return 1;
-    int ret = is_user_process(ts);
-    if (ret < 1) return ret;
-
-    void *real_parent = read_field_ptr(ts, CRC_TASK_STRUCT_REAL_PARENT);
-    if (real_parent == NULL) return -1;
-
-    u32 ppid = -1;
-    if (read_field(real_parent, CRC_TASK_STRUCT_TGID, &ppid, sizeof(ppid)) < 0) return -1;
-
-    // luid could either be in ts->loginuid OR ts->audit->luid. In
-    // most systems it will be in ts->loginuid so let's try that
-    // first.
-    u32 luid = -1;
-    u64 offset_key = CRC_TASK_STRUCT_LOGINUID;
-    u32 *offset = (u32 *)bpf_map_lookup_elem(&offsets, &offset_key);
-    if (offset == NULL) {
-        // if any part of ts->audit->loginuid fails emit a warning as
-        // we expect it there when ts->loginuid is not.
-        void *audit = read_field_ptr(ts, CRC_TASK_STRUCT_AUDIT);
-        if (audit == NULL) return -1;
-        if (read_field(audit, CRC_AUDIT_TASK_INFO_LOGINUID, &luid, sizeof(luid)) < 0) return -1;
-    } else {
-        if (bpf_probe_read(&luid, sizeof(luid), ts + *offset) < 0) {
-            error_info_t info = {0};
-            info.offset_crc = offset_key;
-            set_local_warning(PMW_PTR_FIELD_READ, info);
-
-            return -1;
-        }
-    }
-
-    u64 uid_gid = bpf_get_current_uid_gid();
-
-    syscall_info->pid = pid;
-    syscall_info->ppid = ppid;
-    syscall_info->luid = luid;
-    syscall_info->euid = uid_gid >> 32;
-    syscall_info->egid = uid_gid & 0xFFFFFFFF;
-    syscall_info->mono_ns = bpf_ktime_get_ns();
-
-    return 0;
 }
 
 // it's argument should be a pointer to a file
@@ -964,25 +757,7 @@ int kretprobe__ret_sys_vfork(struct pt_regs *ctx)
 SEC("kprobe/sys_unshare_4_8")
 int BPF_KPROBE_SYSCALL(kprobe__sys_unshare_4_8, int flags)
 {
-    u64 pid_tgid = bpf_get_current_pid_tgid();
-    incomplete_unshare_t event = {0};
-    event.pid_tgid = pid_tgid;
-    event.unshare_flags = flags;
-
-    int ret = bpf_map_update_elem(&incomplete_unshares, &pid_tgid, &event, BPF_ANY);
-    if (ret < 0)
-    {
-        process_message_t pm = {0};
-        pm.type = PM_WARNING;
-        pm.u.warning_info.pid_tgid = pid_tgid;
-        pm.u.warning_info.message_type = PM_UNSHARE;
-        pm.u.warning_info.code = PMW_UPDATE_MAP_ERROR;
-        pm.u.warning_info.info.err = ret;
-
-        push_message(ctx, &pm);
-
-        return 0;
-    }
+    enter_unshare(ctx, flags);
 
     return 0;
 }
@@ -990,42 +765,8 @@ int BPF_KPROBE_SYSCALL(kprobe__sys_unshare_4_8, int flags)
 SEC("kretprobe/ret_sys_unshare")
 int kretprobe__ret_sys_unshare(struct pt_regs *ctx)
 {
-    u64 pid_tgid = bpf_get_current_pid_tgid();
     process_message_t pm = {0};
-    load_event(incomplete_unshares, pid_tgid, incomplete_unshare_t);
-
-    int retcode = (int)PT_REGS_RC(ctx);
-    if (retcode < 0)
-        goto Done;
-
-    void *ts = (void *)bpf_get_current_task();
-
-    int ret = fill_syscall(&pm.u.syscall_info, ts, pid_tgid >> 32);
-    if (ret > 0) return 0;
-    if (ret < 0) goto EmitWarning;
-
-    pm.type = PM_UNSHARE;
-    pm.u.syscall_info.data.unshare_flags = event.unshare_flags;
-    pm.u.syscall_info.retcode = retcode;
-
-    push_message(ctx, &pm);
-    goto Done;
-
- EventMismatch:;
-    error_info_t info = {0};
-    info.stored_pid_tgid = event.pid_tgid;
-    set_local_warning(PMW_PID_TGID_MISMATCH, info);
-
- EmitWarning:;
-    push_warning(ctx, &pm, PM_UNSHARE);
-
- Done:;
-    // only delete at the very end so the event pointer above is valid
-    // for the duration of this function.
-    bpf_map_delete_elem(&incomplete_unshares, &pid_tgid);
-    return 0;
-
- NoEvent:;
+    exit_unshare(ctx, &pm);
     return 0;
 }
